+++ cli/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "cli"
+version = "0.1.0"
+edition.workspace = true
+description.workspace = true
+homepage.workspace = true
+
+[dependencies]
+core = { path = "../core" }
+
+clap = { version = "4.0.22", features = ["derive"] }
+serde = { version = "1.0.147", features = ["derive"] }
+serde_json = "1.0"
