+++ rest_api/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "rest_api"
+version = "0.1.0"
+edition.workspace = true
+description.workspace = true
+homepage.workspace = true
+
+[dependencies]
