--- src/bin/beerus-api.rs
@@ -1,3 +1,0 @@
-fn main() {
-    println!("Running Beerus API");
-}
