--- tests/test_beerus.rs
@@ -1,6 +1,0 @@
-#[cfg(test)]
-mod tests {
-
-    #[test]
-    fn test_hakai() {}
-}
